Fix MPS-25058 Rebuild idea project in presence of test folders leads to erasing generated sources after generation

+++ IdeaPlugin/mps-core/jps-plugin/src/jetbrains/mps/jps/build/MPSModuleLevelBuilder.java
@@ -83,7 +83,7 @@ public class MPSModuleLevelBuilder extends ModuleLevelBuilder {
   // the thing is: out source gen dir is per module, but our builder is called per module chunk. For isntance,
   // a module can be compiled in two goes: 1st chunk - module's sources, 2nd - module tests. Without
   // keeping track we would erase the sources that were generated on the previous step.
-  private Set<JpsModule> cleanedGenSources;
+  private Set<JpsModule> genSourcesNotToClean;
 
   protected MPSModuleLevelBuilder() {
     super(BuilderCategory.SOURCE_GENERATOR);
@@ -97,7 +97,7 @@ public class MPSModuleLevelBuilder extends ModuleLevelBuilder {
 
   @Override
   public void buildStarted(final CompileContext context) {
-    cleanedGenSources = new HashSet<JpsModule>();
+    genSourcesNotToClean = new HashSet<JpsModule>();
     context.addBuildListener(new BuildListener() {
       @Override
       public void filesGenerated(Collection<Pair<String, String>> paths) {
@@ -112,7 +112,7 @@ public class MPSModuleLevelBuilder extends ModuleLevelBuilder {
 
   @Override
   public void buildFinished(CompileContext context) {
-    cleanedGenSources = null;
+    genSourcesNotToClean = null;
     Collection<String> filesToRefresh = refreshComponent.getFilesToRefresh();
     if (!filesToRefresh.isEmpty()) {
       for (String file : filesToRefresh) {
@@ -147,9 +147,6 @@ public class MPSModuleLevelBuilder extends ModuleLevelBuilder {
         continue;
       }
       File outputDir = new File(extension.getConfiguration().getGeneratorOutputPath());
-      if (!outputDir.exists()) {
-        continue;
-      }
       // check that in case it's a module source root then it's marked as generated, only detele of it's true
       Set<File> sourceRootsToKeep = untouchableSourceRoots(compileContext, jpsModule, moduleChunk.getTargets());
 
@@ -161,8 +158,7 @@ public class MPSModuleLevelBuilder extends ModuleLevelBuilder {
         // so in this case it is safe to delete such root
         if (JpsPathUtil.isUnder(sourceRootsToKeep, outputDir)) {
           okToDelete = false;
-        }
-        else {
+        } else {
           final Set<File> _outRoot = Collections.singleton(outputDir);
           for (File srcRoot : sourceRootsToKeep) {
             if (JpsPathUtil.isUnder(_outRoot, srcRoot)) {
@@ -177,14 +173,20 @@ public class MPSModuleLevelBuilder extends ModuleLevelBuilder {
         LOG.warn("Not cleaning generator output path "
           + outputDir.getPath()
           + " because user files may be there. Either mark it as generated or exclude from module");
+        synchronized (this) {
+          genSourcesNotToClean.add(jpsModule);
+        }
         return;
       }
 
-      synchronized (cleanedGenSources) {
-        if (cleanedGenSources.contains(jpsModule)) {
+      synchronized (this) {
+        if (genSourcesNotToClean.contains(jpsModule)) {
           continue;
         }
-        cleanedGenSources.add(jpsModule);
+        genSourcesNotToClean.add(jpsModule);
+      }
+      if (!outputDir.exists()) {
+        continue;
       }
       List<String> deleted = new ArrayList<String>();
       BuildOperations.deleteRecursively(extension.getConfiguration().getGeneratorOutputPath(), deleted, null);
